remove duplicate test, move code to correct location
